+++ src/Tomat.FNB.TMOD/TmodFileEntry.cs
@@ -13,9 +13,9 @@ namespace Tomat.FNB.TMOD;
 /// <param name="CompressedLength">The compressed file length.</param>
 /// <param name="Data">The optionally compressed file data.</param>
 public readonly record struct TmodFileEntry(
-    string           Path,
-    int              Offset,
-    int              Length,
-    int              CompressedLength,
+    string     Path,
+    long       Offset,
+    long       Length,
+    long       CompressedLength,
     IDataView? Data
 );
